refactor(db-types): move augmented Database + AgicashDb rows into @agicash/db-types
Move agicash-db/database.ts (the MergeDeep-augmented Database with wallet RPC
return/composite types, the AgicashDb* row types, and the isCashuAccount/isSparkAccount
guards) plus the 2 pure-zod account-detail json-models it needs into @agicash/db-types,
so the SDK's repositories can consume the DB type layer without importing from the web app.

- database.ts + json-models/{cashu,spark}-account-details-db-data.ts -> packages/db-types/src/
  (byte-identical git mv)
- package index re-exports the AUGMENTED Database (from ./database) + the generated helper
  types (Json/Tables/... from ./database.types) WITHOUT the generated Database, to avoid a
  name clash
- old agicash-db/database.ts + the 2 account-detail json-models kept as transitional
  re-export shims (35 + 8 importers untouched)
- db-types gains deps zod, @supabase/supabase-js, type-fest (latter two hoisted to the root
  catalog, now shared by web + db-types)

Phase 0 of the wallet-sdk extraction. typecheck, biome, tests (web 119, utils 14) and full
build all green; zero web behavior change. The 11 remaining json-models stay in the app for
now (they depend on ~/lib/cashu ProofSchema / the send feature — handled in the cashu lib-ify
and send phases).
