+++ apps/web-wallet/app/features/agicash-db/json-models/spark-account-details-db-data.ts
@@ -1,15 +1,2 @@
-import { z } from 'zod/mini';
-
-export const SparkAccountDetailsDbDataSchema = z.object({
-  /**
-   * Network of the Spark account.
-   * Based on the Breez SDK network type (stored as uppercase in the DB).
-   */
-  network: z.enum(['MAINNET', 'REGTEST']),
-});
-
-export type SparkAccountDetailsDbData = z.infer<
-  typeof SparkAccountDetailsDbDataSchema
->;
-
-export type SparkNetwork = SparkAccountDetailsDbData['network'];
+// Transitional re-export — moved to @agicash/db-types; removed in the import-cleanup PR.
+export * from '@agicash/db-types/json-models/spark-account-details-db-data';
