+++ apps/web-wallet/package.json
@@ -46,7 +46,7 @@
     "@scure/bip39": "1.6.0",
     "@sentry/profiling-node": "10.42.0",
     "@sentry/react-router": "10.42.0",
-    "@supabase/supabase-js": "2.95.2",
+    "@supabase/supabase-js": "catalog:",
     "@tanstack/react-query": "5.90.20",
     "@tanstack/react-query-devtools": "5.91.3",
     "@vercel/analytics": "1.5.0",
@@ -91,7 +91,7 @@
     "npm-run-all": "4.1.5",
     "tailwindcss": "4.1.18",
     "tsx": "4.21.0",
-    "type-fest": "5.4.3",
+    "type-fest": "catalog:",
     "typescript": "catalog:",
     "vite": "7.3.1",
     "vite-plugin-devtools-json": "1.0.0",
