+++ package.json
@@ -8,11 +8,13 @@
       "@agicash/opensecret": "0.1.0",
       "@stablelib/base64": "2.0.1",
       "@stablelib/chacha20poly1305": "2.0.1",
+      "@supabase/supabase-js": "2.95.2",
       "@types/big.js": "6.2.2",
       "@types/bun": "1.3.11",
       "big.js": "7.0.1",
       "dotenv": "16.4.7",
       "jwt-encode": "1.0.1",
+      "type-fest": "5.4.3",
       "typescript": "5.9.3",
       "zod": "4.3.6"
     }
