+++ packages/db-types/src/json-models/spark-account-details-db-data.ts
@@ -1,0 +1,15 @@
+import { z } from 'zod/mini';
+
+export const SparkAccountDetailsDbDataSchema = z.object({
+  /**
+   * Network of the Spark account.
+   * Based on the Breez SDK network type (stored as uppercase in the DB).
+   */
+  network: z.enum(['MAINNET', 'REGTEST']),
+});
+
+export type SparkAccountDetailsDbData = z.infer<
+  typeof SparkAccountDetailsDbDataSchema
+>;
+
+export type SparkNetwork = SparkAccountDetailsDbData['network'];
